+++ CHANGELOG.md
@@ -12,7 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
--
+- Updated neptune logger [#1432](https://github.com/catalyst-team/catalyst/pull/1411)
 
 ### Removed